[DP-229] feat: 채팅 header, inputbar 레이아웃 조정 #184
