+++ src/app/(feature)/chat/[chatRoomId]/layout.tsx
@@ -1,5 +1,5 @@
-import { Fragment, ReactNode } from "react";
+import { ReactNode } from "react";
 
 export default function ChatDetailLayout({ children }: { children: ReactNode }) {
-  return <Fragment>{children}</Fragment>;
+  return <div className="relative h-full w-full overflow-x-hidden bg-white">{children}</div>;
 }
