+++ src/components/common/header/AppHeader.tsx
@@ -22,8 +22,9 @@ export default function AppHeader({ children }: PropsWithChildren<AppHeaderProps
         return pathLen === 2 ? (!isVisible ? "base" : "normal") : "detail";
       case "map":
       case "mypage":
-      case "chat":
         return pathLen === 2 ? "normal" : "detail";
+      case "chat":
+        return pathLen === 2 ? "normal" : "hidden";
       default:
         return "normal";
     }
@@ -78,6 +79,8 @@ export default function AppHeader({ children }: PropsWithChildren<AppHeaderProps
             {children}
           </>
         );
+      case "hidden":
+        return null;
     }
   }, [pathVariant]);
   return (
